requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.27
requests==2.26.0